requirements.txt
# Core Requirements
claude-code-sdk>=0.0.14
anyio>=4.0.0
python-dotenv>=1.0.0

# Utilities
pyyaml>=6.0
structlog>=24.0.0
aiofiles>=23.0.0